Replace TypecheckFailed with UnexpectedType error
Most type issues trigger UnexpectedType which in turn becomes an exception that
can be caught and handled. Writes to typed variables trigger TypecheckFailed
instead, which remains a vm-error. That's odd from a design perspective (what
makes writes so magical that they can't be handled by user code vs. any other
type issues?), and it also makes error reporting much poorer when one is
encountered (although that can be fixed separately and the error reporting was
just the trigger for the investigation).

Moves some test cases around out of vm-lib and into aria code as a result