Allow defining index[] operators with arbitrary argument counts
This is not *exactly* right, as it's not valid to define

operator[]() -> i.e. with 0 arguments; why? because the call site cannot be
foo[] with 0 arguments - but that needs fixed at the call site, not at the
definition site. If you have a use for for foo[] and foo[] = n by all
means, go for it!