Using nanobench as a fast way to benchmark .aria files execution and fast check performance variations (#287)
As I said in other PR, I wanted to do a fast prototype of a benchmark
framework to be able to do two things:
- Bench specifics inside the Rust codebase (already in the master
branch)
- Fast bench different .aria files inside a folder to check regression
or improvements in the overall .aria file interpretation time

The idea I propose here is the use of nanobench to benchmark the .aria
files (right now I'm pointing to the `test/` dir). The 424 files
benchmarked take 3 min (2 min really but the http request take a lot of
time) in my machine.

This is a draft and because of that i would like to push it to a new
branch, so everyone can test it and make suggestions. Right now, to
execute the benchmarks using nanobench, just run th `p` script with no
args.

```bash
./p
```

Tell me your thoughts and if you think something like this is worth it
in the repo