Treat hex literals as unsigned bit patterns
When a 64-bit integer literal is being parsed, assume decimal means a number
but hex and bin literals mean "here's some bits, blast them in an Int"

This allows writing large constants (e.g. as used in hashing) without having
to resort to a less obvious decimal notation. Incidentally, do this in a way
that also transforms a compiler panic into an orderly error

Fixes #294