Add a sample program
This is the equivalent of add_license_marker.sh in Aria

It is not intended for use in CI, since Aria may not be built yet when the
pipeline runs. But it is a good sample of a small yet useful Aria program
and contributors are welcome to run it locally vs. the bash version if they
so desire